requirements.txt
ag2
python-dotenv
pydantic